+++ import_export/templatetags/import_export_tags.py
@@ -1,4 +1,5 @@
 from diff_match_patch import diff_match_patch
+
 from django import template
 
 register = template.Library()